@@ -1,6 +1,6 @@
 {
-	"name": "pvc/testing-traits",
-	"description": "Traits useful for testing",
+	"name": "pvc/testingutils",
+	"description": "utilities for testing",
 	"type": "project",
 	"authors": [
 		{
@@ -10,21 +10,22 @@
 	],
 	"license": "MIT",
 	"config": {
-		"php": "^7.4"
+		"php": "^8.1"
 	},
 	"autoload": {
 		"psr-4": {
-			"pvc\\testingTraits\\": "src",
-			"tests\\": "tests"
+			"pvc\\testingutils\\": "src",
+			"pvcTests\\testingutils\\": "tests"
 		}
 	},
 	"require": {
 		"ext-mbstring": "*"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^8",
-		"mockery/mockery": "dev-master",
-		"squizlabs/php_codesniffer": "^3.5",
-		"phpstan/phpstan": "^0.12.25"
+		"squizlabs/php_codesniffer": "*",
+		"friendsofphp/php-cs-fixer": "*",
+		"phpunit/phpunit": "^9",
+		"mockery/mockery": "*",
+		"phpstan/phpstan": "*"
 	}
 }
