@@ -1,39 +1,37 @@
 <?php
+
+declare (strict_types=1);
 /**
- * @package: pvc
  * @author: Doug Wilbourne (dougwilbourne@gmail.com)
- * @version: 1.0
  */
 
-namespace tests\mockery;
+namespace pvcTests\testingutils\testingTraits\phpunit;
 
-use Mockery;
+use Iterator;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use pvc\testingTraits\mockery\MockeryIteratorTrait;
+use pvc\testingutils\testingTraits\phpunit\PhpunitIteratorTrait;
 
-/**
- * Class MockeryIteratorTraitTest
- * @package tests
- */
-class MockeryIteratorTraitTest extends Testcase
+class PhpunitIteratorTraitTest extends TestCase
 {
+    use PhpunitIteratorTrait;
 
-    use MockeryIteratorTrait;
-
-    /** @phpstan-ignore-next-line */
-    protected $mock;
+    protected MockObject|Iterator $mock;
     protected array $testArray;
 
-    public function setUp() : void
+    public function setUp(): void
     {
-        $this->mock = Mockery::mock('\StdClass', '\Iterator');
+        $this->mock = $this->createMock(Iterator::class);
         $this->testArray = array('foo', 'bar', 'baz');
         $this->mock = $this->mockIterator($this->mock, $this->testArray);
     }
 
-    public function testIteration() : void
+    /**
+     * testIteration
+     * @covers PhpunitIteratorTrait::mockIterator
+     */
+    public function testIteration(): void
     {
-
         // if you have a mocked object that needs to implement the iterator interface,
         // you can use the function mockIterator to add the proper expectations
 
@@ -49,9 +47,12 @@ class MockeryIteratorTraitTest extends Testcase
         $this->assertSame($this->testArray, $arrayValues);
     }
 
-    public function testIteratorMethods() : void
+    /**
+     * testIteratorMethods
+     * @covers PhpunitIteratorTrait::mockIterator
+     */
+    public function testIteratorMethods(): void
     {
-
         $this->assertTrue($this->mock->valid());
         $this->assertEquals(0, $this->mock->key());
         $this->assertEquals('foo', $this->mock->current());
@@ -63,8 +64,6 @@ class MockeryIteratorTraitTest extends Testcase
         $this->mock->next();
         $this->assertEquals(2, $this->mock->key());
         $this->assertEquals('baz', $this->mock->current());
-
-        $this->assertEquals(3, $this->mock->count());
 
         $this->mock->next();
         $this->assertFalse($this->mock->valid());
